Added Camera Kernel + Row, Col Color Shader
Added a script that applies a convolution kernel to a 2D texture camera render conversion. This is inefficient, so it likely won't be worked on or expanded upon further.

I also added shaders that split the screen colours into three tints. There's one to do it by rows, and another to do it by columns.