Added Mix Shader + Find Space World Units in Grid + String Helper Fixes
Added a shader that mixes the rendered image with a texture, with the user determining how strong the mix should be with lerp.

Added a function that allows the user to find a space in a grid in world units. This was adapted from another project, so it should work.

Also took out unneeded includes in the string helper.